Integrate Justin's community URL changes with the new site split changes
