+++ console/test/rails_app/config/environments/test.rb
@@ -38,8 +38,7 @@ RailsApp::Application.configure do
 
   Console.configure('~/.openshift/console.conf') do |c|
     c.security_controller = 'Console::Auth::Basic' # test cases depend on basic authentication today
-    #c.community_url = 'https://www.openshift.com'
-    c.community_url = 'https://openshift.redhat.com/community'
+    c.community_url = 'https://www.openshift.com'
   end
 end
